tools/requirements.txt
# Hyperliquid Signer Bridge Dependencies
eth-account>=0.11.0
hyperliquid-python-sdk>=0.4.0

# Additional dependencies for trading strategies
pyzmq>=25.0.0
numpy>=1.24.0
